Use shorter timestamps for Michelle
The exec filenames easily become too long for Michelle.  Therefore
omit the microseconds and then add three digits chosen to ensure
a unique filename.  For other instruments, also use this procedure
but include milliseconds in the timestamp part of the name.
This matches the UKIRT translator's exec file naming scheme.